[lib] Move `json-schema-compare` to lib
